move python test source to .py files
